* main: (298 commits)
  Cleanup.
  Use AsCommand attribute.
  Raise elasticsearch to 7.17.2
  Add loggly.
  And they’re attributes now.
  This is called EnumType now.
  Update maltehuebner/impress-bundle
  Update recipe for symfony/routing
  Fix symfony versions to ^6.2
  Fix param converter attributes.
  Remove sensio/framework-extra-bundle
  Cleanup dependencies.
  Deprecate getDoctrine() call.
  Fix missing getDoctrine() call.
  Fix main request.
  Fix all the parameter annotations.
  Update openapi imports.
  Remove composer/package-versions-deprecated.
  Remove fahrverbote from menu.
  Remove fahrverbote from city page.
  ...

# Conflicts:
#	docker-compose.yml

+++ src/Command/PurgeDataCommand.php
@@ -4,7 +4,8 @@ namespace App\Command;
 
 use App\DataPurger\DataPurgerInterface;
 use App\Provider\ProviderListInterface;
-use App\Util\DateTimeUtil;
+use Carbon\CarbonImmutable;
+use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
@@ -12,22 +13,20 @@ use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
 
+#[AsCommand(
+    name: 'luft:purge-data',
+    description: 'Purge data from elasticsearch'
+)]
 class PurgeDataCommand extends Command
 {
-    protected ProviderListInterface $providerList;
-    protected DataPurgerInterface $dataPurger;
-
-    public function __construct(?string $name = null, ProviderListInterface $providerList, DataPurgerInterface $dataPurger)
+    public function __construct(protected ProviderListInterface $providerList, protected DataPurgerInterface $dataPurger)
     {
-        $this->providerList = $providerList;
-        $this->dataPurger = $dataPurger;
-
-        parent::__construct($name);
+        parent::__construct();
     }
 
     protected function configure(): void
     {
-        $this->setName('luft:purge-data')
+        $this
             ->addArgument('days', InputArgument::REQUIRED, 'Specify number of days. Data older than this value will be purged.')
             ->addArgument('provider', InputArgument::OPTIONAL, 'Optional: Specify provider to purge.')
             ->addOption('with-tags', 'wt', InputOption::VALUE_NONE, 'Also delete tagged data')
@@ -49,16 +48,28 @@ class PurgeDataCommand extends Command
         }
 
         $interval = new \DateInterval(sprintf('P%dD', $input->getArgument('days')));
-        $untilDateTime = DateTimeUtil::getDayEndDateTime((new \DateTime())->sub($interval));
+        $untilDateTime = CarbonImmutable::now()->sub($interval);
+
+        $provider = $this->providerList->getProvider($input->getArgument('provider'));
+
+        $dataList = $this->registry->getRepository(Data::class)->findInInterval(null, $untilDateTime, $provider);
+
+        $helper = $this->getHelper('question');
+        $question = new ConfirmationQuestion(sprintf('Purge <info>%d</info> values from <comment>%s</comment> before <info>%s</info>? [no] ', is_countable($dataList) ? count($dataList) : 0, $provider::class, $untilDateTime->format('Y-m-d H:i:s')), false);
+
+        if (!$helper->ask($input, $output, $question)) {
+            return 1;
+        }
 
         $counter = $this->dataPurger->purgeData($untilDateTime, $provider, $input->getOption('with-tags'));
 
         if ($provider) {
-            $io->success(sprintf('Purged %d values from %s.', $counter, get_class($provider)));
+            $io->success(sprintf('Purged %d values from %s.', $counter, $provider::class));
         } else {
             $io->success(sprintf('Purged %d values.', $counter));
         }
 
+        //return Command::SUCCESS;
         return 0;
     }
 }
